backend/pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic<2.0.0,>=1.14.0",
    # Pin bcrypt until passlib supports the latest
    "bcrypt==4.0.1",
    "fastapi[standard]<1.0.0,>=0.115.6",
    "jinja2<4.0.0,>=3.1.4",
    "orjson<4.0.0,>=3.10.12",
    "passlib[bcrypt]<2.0.0,>=1.7.4",
    "psycopg2-binary<3.0.0,>=2.9.10",
    "pydantic-settings<3.0.0,>=2.7.0",
    "pyjwt<3.0.0,>=2.10.1",
    "sqlalchemy-utils<1.0.0,>=0.41.2",
    "sqlmodel<1.0.0,>=0.0.22",
    "tenacity<10.0.0,>=9.0.0",
]

[dependency-groups]
dev = [
    "coverage<8.0.0,>=7.6.9",
    "mypy<2.0.0,>=1.13.0",
    "pre-commit<5.0.0,>=4.0.1",
    "pytest<9.0.0,>=8.3.4",
    "ruff<1.0.0,>=0.8.3",
    "types-passlib<2.0.0,>=1.7.7.20240819",
    "types-psycopg2<3.0.0,>=2.9.21.20241019",
    "types-pytz<2025.0.0,>=2024.2.0.20241221",
]

# MYPY section
[tool.mypy]
plugins = [
    "pydantic.mypy",
]
strict = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_optional = true
exclude = [".env", ".venv"]

# RUFF section
[tool.ruff]
line-length = 88
extend-exclude = [".env", ".venv"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]
